fix stupid error with signed vs. unsigned
In the case that the EclFileName parameter did not contain a '/', this
code led to a segmentation fault. things happens if you set the
warning level of your compiler too high and "fix" the false positive
warnings incorrectly.

+++ applications/ebos/eclgridmanager.hh
@@ -164,7 +164,7 @@ public:
 
         // compute the base name of the input file name
         const char directorySeparator = '/';
-        unsigned i;
+        long int i;
         for (i = fileName.size(); i >= 0; -- i)
             if (fileName[i] == directorySeparator)
                 break;
